pyproject.toml
[project]
name = "downloader-cli"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "tadeasf", email = "taddy.fort@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "typer>=0.12.5",
    "pyaml>=24.7.0",
    "toml>=0.10.2",
    "prompt-toolkit>=3.0.47",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/downloader_cli"]

[project.scripts]
downloader = "downloader_cli.main:app"
